4717 - Moved submitter to ingest core module
