+++ java/ingest/ingest-batcher-core/src/main/java/sleeper/ingest/batcher/core/IngestBatcherSubmitRequestSerDe.java
@@ -14,14 +14,19 @@
  * limitations under the License.
  */
 
-package sleeper.ingest.batcher.submitter;
+package sleeper.ingest.batcher.core;
 
-import java.util.List;
-import java.util.stream.Collectors;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 
-public record IngestBatcherSubmitRequest(List<String> files, String tableName) {
+public class IngestBatcherSubmitRequestSerDe {
+    private final Gson gson = new GsonBuilder().create();
 
-    public IngestBatcherSubmitRequest(String bucketName, List<String> keys, String tableName) {
-        this(keys.stream().map(key -> bucketName + "/" + key).collect(Collectors.toList()), tableName);
+    public IngestBatcherSubmitRequest fromJson(String json) {
+        return gson.fromJson(json, IngestBatcherSubmitRequest.class);
+    }
+
+    public String toJson(IngestBatcherSubmitRequest request) {
+        return gson.toJson(request);
     }
 }
