+++ java/ingest/ingest-batcher-submitter/src/main/java/sleeper/ingest/batcher/submitter/IngestBatcherSubmitter.java
@@ -25,6 +25,7 @@ import sleeper.core.table.TableNotFoundException;
 import sleeper.core.table.TableStatus;
 import sleeper.core.util.NumberFormatUtils;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
+import sleeper.ingest.batcher.core.IngestBatcherSubmitRequest;
 import sleeper.ingest.batcher.core.IngestBatcherTrackedFile;
 import sleeper.parquet.utils.HadoopPathUtils;
